pyproject.toml
[project]
name = "anomaly-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "plotly>=6.6.0",
    "polars>=1.39.3",
    "pyarrow>=23.0.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.3",
]
